backend/requirements.txt
# Core Framework
fastapi==0.125.0
mangum==0.19.0

# Auth - JWT verification for Cognito (requires cryptography for RS256)
PyJWT==2.10.1
cryptography==44.0.0

# HTTP Client - trigger deployment
httpx==0.28.1

# AWS SDK (Async) - CRITICAL: Keep these versions or newer to avoid ClientArgsCreator errors
# DO NOT DOWNGRADE: These versions fix compute_endpoint_resolver_builtin_defaults() missing args
aioboto3>=15.5.0
aiobotocore>=2.25.1
botocore>=1.35.67

# Data Validation
pydantic==2.12.5
pydantic-settings==2.6.1

# Performance & Caching
cachetools==6.2.4
